requirements.txt
streamlit
pandas
openpyxl
gspread
gspread-dataframe
google-auth
google-api-python-client